Module_6_Multi_Agents/advance-customer-support-agent-feature-A2A-MCP-ADK/requirements.txt
python-dotenv>=0.9.9
google-adk>=1.19.0
toolbox-core>=0.5.3
mem0ai>=0.1.118
psycopg2>=2.9.11
tabulate>=0.9.0

# A2A / Security pipeline dependencies
google-cloud-dlp>=3.28.0
google-cloud-secret-manager>=2.23.2
google-auth>=2.38.0
aiohttp>=3.9.0
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0
loguru>=0.7.0
requests>=2.31.0